requirements.txt
flask
gunicorn
numpy
pillow
opencv-python